EmblaCarousel.js Updated
