+++ helpers/CarouselImages.js
@@ -1,0 +1,23 @@
+const route = '/Images/Banner/'
+    const slides = [
+        {
+            name: "banner-01",
+            description: "description #1",
+            url: `${route}banner-01.webp`
+        },
+        {
+            name: "banner-02",
+            description: "description #2",
+            url: `${route}banner-02.webp`
+       
+        },
+        {
+            name: "banner-03",
+            description: "description #3",
+            url: `${route}banner-03.webp`
+       
+        }
+      ]
+ 
+
+export default slides;
