+++ styles/EmblaCarousel.module.css
@@ -1,0 +1,65 @@
+.embla {
+    position: relative;
+  
+    max-width: 100%;
+    margin-left: auto;
+    margin-right: auto;
+  
+  
+  }
+  
+  .embla__viewport {
+    overflow: hidden;
+    width: 100%;
+  
+  }
+  
+  .embla__viewport.is-draggable {
+    cursor: move;
+    cursor: grab;
+  }
+  
+  .embla__viewport.is-dragging {
+    cursor: grabbing;
+  }
+  
+  .embla__container {
+    display: flex;
+    user-select: none;
+    -webkit-touch-callout: none;
+    -khtml-user-select: none;
+    -webkit-tap-highlight-color: transparent;
+    margin-left: -10px;
+    height: 100%;
+  }
+  
+  .embla__slide {
+    position: relative;
+    min-width: 100%;
+    padding-left: 10px;
+  }
+  
+  .embla__slide__inner {
+    position: relative;
+    overflow: hidden;
+    
+  }
+  
+  .embla__slide__img {
+    position: relative;
+    display: block;
+    top: 50%;
+    left: 50%;
+    width: 100%;
+    min-width: 100%;
+    max-width: none;
+    height: 100%;
+    transform: translate(-50%, 0%);
+  }
+  
+  .embla__dots {
+    display: flex;
+    list-style: none;
+    justify-content: center;
+    padding-top: 10px;
+  }
